+++ src/test/java/com/opensource/TestContainerTest.java
@@ -9,44 +9,14 @@ import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static org.assertj.core.api.Assertions.assertThat;
-@Testcontainers
-public class TestContainerTest {
-    @Container
-    private static final PostgreSQLContainer<?> postgreSQLContainer =
-            new PostgreSQLContainer<>("postgres:latest")
-                    .withDatabaseName("ezhil-dao-unit-test")
-                    .withUsername("ezhil")
-                    .withPassword("password");
 
-    @DynamicPropertySource
-    private static void registerDataSourceProperties(DynamicPropertyRegistry registry){
-        registry.add(
-                "spring.datasource.url",
-                postgreSQLContainer::getJdbcUrl
-        );
-        registry.add(
-                "spring.datasource.username",
-                postgreSQLContainer::getUsername
-        );
-        registry.add(
-                "spring.datasource.password",
-                postgreSQLContainer::getPassword
-        );
-    }
+public class TestContainerTest extends AbstractTestContainer{
+
     @Test
     void canStartPostgresDB() {
        assertThat(postgreSQLContainer.isRunning()).isTrue();
        assertThat(postgreSQLContainer.isCreated()).isTrue();
     }
 
-    @Test
-    void canApplyDBMigrationWithFlyway(){
-        Flyway flyway = Flyway.configure().dataSource(
-                postgreSQLContainer.getJdbcUrl(),
-                postgreSQLContainer.getUsername(),
-                postgreSQLContainer.getPassword()
-        ).load();
-        flyway.migrate();
-        System.out.println();
-    }
+
 }
